+++ CHANGELOG.md
@@ -35,6 +35,7 @@ Table of releases
 - refactored yandex
 - refactored tgos
 - refactored opencage
+- refactored maxmind
 - fixed `base.py: MultipleResultsQuery._catch_errors` which was not returning errors most of the time
 - fixed issue #207: tomtom, making use of new API